+++ docs/changes.rst
@@ -20,6 +20,7 @@ Weblate 5.17.1
 * Track superuser and site-wide team changes in :ref:`audit-log`.
 * Installing :ref:`addon-weblate.discovery.discovery` on a component now suggests additional guided presets detected from the component repository layout.
 * Project website and repository browser URL alerts now show validation errors more clearly.
+* Attribute translations copied by :ref:`addon-weblate.autotranslate.autotranslate` to the add-on user.
 
 .. rubric:: Bug fixes